requirements.txt
numpy
matplotlib
ipykernel
PyYAML